+++ src/json_t.mli
@@ -98,7 +98,8 @@ type ghost_abilities = {
 type game_progress = {
   mutable frame_idx: int;
   steel_sole: steel_sole_progress;
-  mutable by_room: (string * room_progress) list
+  mutable by_room: (string * room_progress) list;
+  mutable keys_found: string list
 }
 
 type save_file = {